Step 6: Replace AST-size extraction with FusionCostModel
Uses egglog's extract_value_with_cost_model API to apply an HBM-traffic-
aware cost model during e-graph extraction, replacing the default AST-
size cost.

FusionCostModel assigns per-constructor costs reflecting relative GPU
execution expense:
  - Leaf nodes (Input/Parameter/Const): 0 (always needed)
  - Pointwise unary/binary: 5
  - Reductions: 10
  - Norms: 15
  - MatMul: 50; FusedMatMulAdd: 45; FusedRmsNormMatMul: 40
  - Attention: 100
  - Conv2d: 80; WinogradConv2d: 60

Fused ops are cheaper than their unfused equivalents (e.g., MatMul +
Add = 55 > FusedMatMulAdd = 45), so the extractor prefers fused forms.

For the current rewrite rules, the cost model produces identical
extraction results to AST-size (both prefer fewer/fused ops). The
infrastructure enables future shape-aware costs and new rewrites where
the cost trade-off isn't AST-size-obvious (e.g., Winograd where the
fused form has MORE nodes but LESS compute).

Implementation: after equality saturation, for each graph output,
eval_expr gets the egglog Value, then extract_value_with_cost_model
picks the minimum-cost equivalent. Fallback to legacy text-based
extraction if the programmatic path fails.

All 206 tests pass.

+++ tests/training_correctness.rs
@@ -94,7 +94,7 @@ fn smollm2_training_loss_decreases() {
     let token_ids: Vec<u32> = (0..seq_len as u32).map(|i| i % vocab as u32).collect();
     let mut labels = vec![0.0f32; seq_len * vocab];
     for i in 0..seq_len {
-        let target = ((i + 1) % vocab) as usize;
+        let target = (i + 1) % vocab;
         labels[i * vocab + target] = 1.0;
     }
